requirements.txt
certifi==2018.4.16
dj-database-url==0.5.0
Django==2.0.6
django-bootstrap3==10.0.1
django-braces==1.13.0
djangorestframework==3.8.2
gunicorn==19.8.1
pipenv==2018.5.18
psycopg2==2.7.4
pytz==2018.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
